refactor: 토론 카드 리스트 컴포넌트 결합
