+++ src/pages/home/api/query.ts
@@ -1,7 +1,7 @@
 import { useQuery } from "@tanstack/react-query";
 import { axios } from "../../../shared/axiosApi";
 import { ServerResponse } from "../../../shared/axiosApi/model/axiosInstance";
-import { DebateMode, RoomCardType } from "../../../shared/type";
+import { RoomCardType } from "../../../shared/type";
 
 export type GetDebateRoomsResponse = RoomCardType[]
 
@@ -13,32 +13,25 @@ export type QueryString = {
   size?: string;
 }
 
-export const getStartedDebate = (
-  mode: DebateMode
-): Promise<ServerResponse<GetDebateRoomsResponse>> => axios.get(`/chat/chatrooms/${mode}`, {
-  params: { sort: "participants", order: "desc", status: "STARTED", page:"0", size: "5" } as QueryString
+export const getDebateRooms = (
+  status: string
+): Promise<ServerResponse<GetDebateRoomsResponse>> => axios.get(`/chat/chatrooms/찬반`, {
+  params: { sort: "participants", order: "desc", status, page:"0", size: "5" } as QueryString
 });
 
 export const useGetStartedDebateQuery = () => {
   return useQuery({
     queryKey: ["getStartedDebate"],
-    queryFn: async () => getStartedDebate("찬반"),
+    queryFn: async () => getDebateRooms("STARTED"),
     refetchInterval: 10000,
     retry: 1,
   });
 }
 
-
-export const getCreatedDebate = (
-  mode: DebateMode
-): Promise<ServerResponse<GetDebateRoomsResponse>> => axios.get(`/chat/chatrooms/${mode}`, {
-  params: { sort: "participants", order: "desc", status: "CREATED", page:"0", size: "5" } as QueryString
-});
-
 export const useGetCreatedDebateQuery = () => {
   return useQuery({
-    queryKey: ["getStartedDebate"],
-    queryFn: async () => getCreatedDebate("찬반"),
+    queryKey: ["getCreatedDebate"],
+    queryFn: async () => getDebateRooms("CREATED"),
     refetchInterval: 10000,
     retry: 1,
   });
